+++ src/User/Infrastructure/Security/JwtCreatedListener.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\User\Infrastructure\Security;
+
+use App\User\Domain\Entity\User;
+use Lexik\Bundle\JWTAuthenticationBundle\Event\JWTCreatedEvent;
+
+class JwtCreatedListener
+{
+    public function onJWTCreated(JWTCreatedEvent $event): void
+    {
+        $user = $event->getUser();
+
+        if (!$user instanceof User) {
+            return;
+        }
+
+        $payload = $event->getData();
+        $payload['userId'] = $user->getId();
+        $event->setData($payload);
+    }
+}